requirements.txt
fastapi==0.104.1
streamlit==1.28.1
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.4
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
lxml==4.9.3
fake-useragent==1.4.0
webdriver-manager==4.0.1
aiofiles==23.2.1
python-multipart==0.0.6 